cosmetic fixes in config, UI, and stats.cvs
- allow fraction for -abort_time
- add header to stats.cvs
- fix perf widget in kafl_gui.py

+++ kAFL-Fuzzer/common/config.py
@@ -151,7 +151,7 @@ def add_args_fuzzer(parser):
     parser.add_argument('-cpu_offset', metavar='<n>', help="offset CPU pinning to start at <N>",
                         type=int, default=0, required=False)
     parser.add_argument('-abort_time', metavar='<n>', help="exit after n hours",
-                        type=int, required=False, default=None)
+                        type=float, required=False, default=None)
     parser.add_argument('-abort_exec', metavar='<n>', help="exit after max executions",
                         type=int, required=False, default=None)
     parser.add_argument('-ts', '--t_soft', required=False, metavar='<s>', help="soft timeout for Qemu execution (in seconds)",
